+++ sympy/printing/latex.py
@@ -16,6 +16,8 @@ from sympy.core.operations import AssocOp
 from sympy.core.sympify import SympifyError
 from sympy.logic.boolalg import true
 
+from sympy.core._print_helpers import LatexPrintable
+
 # sympy.printing imports
 from sympy.printing.precedence import precedence_traditional
 from sympy.printing.printer import Printer
@@ -121,30 +123,7 @@ _between_two_numbers_p = (
 )
 
 
-class LatexPrintable:
-    """
-    The default implementation of latex printing for SymPy classes.
-
-    Mix in this trait into a class to get proper latex printing with IPython.
-    """
-
-    # We don't define _repr_png_ here because it would add a large amount of
-    # data to any notebook containing SymPy expressions, without adding
-    # anything useful to the notebook. It can still enabled manually, e.g.,
-    # for the qtconsole, with init_printing().
-    def _repr_latex_(self):
-        """
-        IPython/Jupyter LaTeX printing
-
-        To change the behavior of this (e.g., pass in some settings to LaTeX),
-        use init_printing(). init_printing() will also enable LaTeX printing
-        for built in numeric types like ints and container types that contain
-        SymPy objects, like lists and dictionaries of expressions.
-        """
-        s = latex(self, mode='plain')
-        return "$\\displaystyle %s$" % s
-
-    _repr_latex_orig = _repr_latex_
+LatexPrintable.__module__ = __name__
 
 
 class LatexPrinter(Printer):
